Update pythontex3.py
This code added "tab support". If user have in pycode environment tab align all lines of code, for example:
\begin{pycode}
	   import numpy as np
	   x=np.linspace(0,10,100)
	   y=np.cos(x)
	   def draw(x,y,param='black'):
		   s=''
		   for i in range(0,len(x)):
# 			   s+=' (%s,%s)'%(x[i],y[i])
		   return str(r'\draw plot[%s] coordinates{%s};'%(param,s))
\end{pycode}
It's really convenient for those who are interested in the beauty of the tabulation code.

Note: this code only for tab (not spaces), and python code in pycode environments required first line as filled of user python code (
as in the example above).

This is still a very weak code, but I want you to fully implement this feature. Sorry, I'm a physicist, not a programmer)